requirements.txt
aioredis==1.2.0
asgiref==3.2.10
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
autopep8==1.5.3
botocore==1.17.60
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.2
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
docutils==0.15.2
future==0.16.0
hiredis==0.2.0
idna==2.7
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.1.0
pathspec==0.5.9
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
python-dateutil==2.8.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
semantic-version==2.5.0
six==1.11.0
sqlparse==0.4.1
termcolor==1.1.0
txaio==18.8.1
urllib3==1.24.3
wcwidth==0.1.9
wrapt==1.10.11
